+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Changelog
 
+### 1.5
+* Added various new output options for shortcode display
+* Updated CSS in `public/css/wp-dispensary-public.css` for improved display in the widgets
+* Added WPD icon to each menu section in dashboard "At A Glance" box
+* Added option to randomize menu item output in widgets
+* Added plugin recommendation for [Dispensary Coupons](https://wordpress.org/plugins/dispensary-coupons/)
+* PHP_CodeSniffer WordPress Coding Standards updates throughout various plugin files
+
 ### 1.4
-* PHP CodeSniffer WordPress Standards updates throughout various plugin files
-* Created the Topicals menu type, with widget, shortcode and WP-API output options
-* Updated the output style of the shortcode item information
-* Added the WP Dispensary icon to each CPT in the admin dashboard for easier recognition when you're in the admin dashboard
+* PHP_CodeSniffer WordPress Coding Standards updates throughout various plugin files
+* Created the Topicals menu type, with Widget, Shortcode and WP-API output options
+* Updated the output style of the Shortcode item information
+* Added the WP Dispensary icon to each CPT in the admin dashboard for easier visual recognition
 * Added `active_plugins` check for "Subtitles" plugin in `admin/wp-dispensary-rest-api.php` API output
 * Added sanitization to $_POST in `admin/post-types/wp-dispensary-metaboxes.php` pre-roll flower selection drop down